requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
future==0.17.1
httplib2==0.13.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
Pillow==6.1.0
pycryptodome==3.9.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
qcloudsms-py==0.1.3
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
XlsxWriter==1.2.1
xlwt==1.3.0
xmltodict==0.12.0
xmltojson==0.2.0
